Debug quantized

+++ pretrain_tuneavideo.py
@@ -361,7 +361,7 @@ def main(
 
             with accelerator.accumulate(unet):
                 # Convert videos to latent space
-                pixel_values = batch["pixel_values"].to(weight_dtype)
+                pixel_values = batch["pixel_values"]  # .to(weight_dtype)
                 video_length = pixel_values.shape[1]
                 pixel_values = rearrange(pixel_values, "b f c h w -> (b f) c h w")
                 latents = vae.encode(pixel_values).latent_dist.sample()
